delivered sub fraction plus calculated sub fraction from time tick value might result in a whole fraction

+++ include/streamclient/subscribedsignal.h
@@ -94,10 +94,8 @@ namespace hbm {
 
 			double m_valueBuffer[m_valueBufferMaxValues];
 
-			uint32_t m_subFraction;
+			uint64_t m_subFraction;
 			uint64_t m_subFractionCollected;
-//			unsigned int m_syncSignalCorrectionCycle;
-//			unsigned int m_syncSignalCorrectionCount;
 		};
 	}
 }
